ocr_checkbox_agent/requirements.txt
# Core PDF processing
PyPDF2==3.0.1
pdf2image==1.16.3
pymupdf==1.23.8

# OCR and image processing
pytesseract==0.3.10
opencv-python==4.8.1.78
Pillow==10.1.0

# Data handling
pandas==2.1.4
openpyxl==3.1.2
numpy==1.26.2

# Utilities
python-dateutil==2.8.2
tqdm==4.66.1
pydantic==2.5.2

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# Logging and configuration
python-dotenv==1.0.0
loguru==0.7.2